Stable Release 1.0.0.0
This is the first stable release of the designer. Build with VB 2008 Express or Studio 2008. 
Included a log writer.
Scriptus now fully functional.
Designer is fully functional.
Dialogues save file integrity at 100%
Scripts save file integrity at 100%

Conditions for running Dialogues are not yet included, and might never be.
Dialogue order logic will have to be done in the Source Code of the Game.

Will be included in the next version of the Engine. Co-Char-Key will not feature a new version though.